requirements.txt
flask>=2.1.2
gevent>=21.12.0
psycopg2-binary>=2.9.3
pydantic>==1.9.1
python-dotenv>=0.20
sqlalchemy>=1.4.39
